Add pools-only model
This commit represents an evolution of the model stripping out many of
the complexities of the KubeCon EU version of the models, and seeing how
much we can represent using only pools of identical devices.

This will be replaced in subsequent commits with models that add back in
individual devices.