Update classes for new model
A class can now either be a set of criteria, or a label selector.

One limitation here is that since we cannot combine the label selector
and a Constraints, we cannot create classes that apply constraints
across another set of classes. So, to do "any 1Gbps SR-IOV NIC", I need
to create a bunch of classes, one for each vendor, that constrains to
1Gbps NICs, and add an appropriate DeviceClass label, and then create
the "any" class based on those labels. That's not *ideal*, but it does
work. The user could also just use a "any SR-IOV NIC" class and put
the constraint directly in the claim, in this case.