dra-evolution: validate CEL expressions
This imports the CEL implementation from DRA in 1.30, updates it to use the
"device" variable name and to support fully qualified attribute names ("model"
and "model.dra.example.com").

Then the parsed objects gets validated, including compilation of the CEL
expressions. This cannot cover runtime errors. The type-safe approach uncovers
errors like comparing a quantity to a string, which is not supported in
Kubernetes because it makes it impossible to do a cost analysis of an
expression.