dra-evolution: harmonize fields
"filter" and "match" now are imperative verbs with a list of items that all
must be true. Each entry is a one-of, which makes it possible to add other
filter or match criteria without them being accidentally ignored.

An individual request can also contain multiple filters or none, which solves
the problem of having to use `device: {}` because the empty list doesn't need
that.