dra-evolution: introduce "claim requirements"
There's only one "claim requirement" right now (match on an attribute) but
others might be added in the future. The advantage of putting them into a
one-of is that old schedulers will not accidentally ignore a hard requirement,
which is what would happen when we added them as a new field.

While at it, related options get grouped together in the class.