dra-evolution: remove multi-inheritance of classes
Inheriting from multiple classes was meant to support use cases where admins
pre-configure "fragments" that users then apply. It's unclear whether that is
actually useful.

If it turns out that it is, then there is now a path towards adding it through
the ClassReference one-of struct: we can add `inherit` into `ClassReference` as
an alternative to `resourceClassName`.