dra-evolution: ResourceClass -> DeviceClass
The concept of a general-purpose ResourceClass which can by itself act as a
replacement for a claim was too confusing and vague. A DeviceClass with
configuration and requirements for a single device is better.

Without a class reference in the claim, the user has to ask for allocation via
a control plane controller in the claim.