dra-evolution: restore minimal support for "classic DRA"
"Classic DRA" has a control plane controller which handles allocation and
deallocation of a claim in cooperation with the scheduler. That is the part
which gets added back to the API.

What doesn't get restored compared to v1.30.0 is:
- immediate allocation
- using a CRD as user-facing API
- the opaque "resource handle" in the allocation result

Drivers which want to support this mode of operation have to use the in-tree
types. The "config" fields support that, albeit perhaps with a slightly less
natural API.

While at it, some occurrences of "resource" get replaced with "device" or
"DRA".